refactor(binance-api): extract convert API to module
Move Binance Convert API methods to a dedicated module:
- Create new convert_api.rs module with ConvertApi class
- Update lib.rs to expose API through a facade method
- Extract all Convert-specific endpoints to the new module
- Make internal methods public for cross-module usage
- Add helper for making public GET requests
