+++ tee-worker/omni-executor/binance-api/src/lib.rs
@@ -1,6 +1,8 @@
+mod convert_api;
 mod error;
 pub mod types;
 
+use convert_api::ConvertApi;
 use error::Error;
 use hmac::{Hmac, Mac};
 use log::error;
@@ -10,17 +12,11 @@ use std::{
 	collections::HashMap,
 	time::{SystemTime, UNIX_EPOCH},
 };
-use types::{
-	AssetInfo, AssetSymbol, CancelOrderLimit, ConvertOrder, ConvertOrderStatus,
-	ConvertTradeHistory, LimitOrder, OpenOrders, PlaceLimitOrderParams, Quote, RequestQuoteParams,
-	TokenPair,
-};
 use url::Url;
 
-const CONVERT_API: &str = "/sapi/v1/convert";
 const MAX_RECV_WINDOW: u32 = 60000;
-const MAX_HISTORY_LIMIT: u16 = 1000;
 
+#[derive(Debug, Clone)]
 pub struct BinanceApi {
 	client: Client,
 	base_url: Url,
@@ -38,157 +34,13 @@ impl BinanceApi {
 		BinanceApi { client, base_url, api_key, api_secret }
 	}
 
-	/// List all convertible token pairs and the tokens’ respective upper/lower limits
-	pub async fn get_exchange_info(
-		&self,
-		from_asset: Option<AssetSymbol>,
-		to_asset: Option<AssetSymbol>,
-	) -> Result<Vec<TokenPair>, Error> {
-		let endpoint = format!("{}/exchangeInfo", CONVERT_API);
-		let mut url = self.base_url.join(&endpoint).unwrap();
-		if let Some(from_asset) = from_asset {
-			url.query_pairs_mut().append_pair("fromAsset", &from_asset);
-		}
-		if let Some(to_asset) = to_asset {
-			url.query_pairs_mut().append_pair("toAsset", &to_asset);
-		}
-		let response = self.client.get(url.as_str()).send().await.map_err(|e| {
-			error!("Error getting exchange info: {}", e);
-			Error::RequestFailed
-		})?;
-		let token_pairs: Vec<TokenPair> = response.json().await.map_err(|e| {
-			error!("Error parsing exchange info: {}", e);
-			Error::ParseResponseFailed
-		})?;
-
-		Ok(token_pairs)
+	/// Create a new ConvertApi instance
+	pub fn convert(&self) -> ConvertApi {
+		ConvertApi::new(self)
 	}
 
-	/// List all supported asset’s precision information
-	pub async fn get_asset_info(&self, recv_window: Option<u32>) -> Result<Vec<AssetInfo>, Error> {
-		let endpoint = format!("{}/assetInfo", CONVERT_API);
-		self.make_signed_request(&endpoint, Method::GET, None, recv_window).await
-	}
-
-	/// Request a quote for the requested token pairs
-	pub async fn get_quote(
-		&self,
-		request_quote_params: RequestQuoteParams,
-		recv_window: Option<u32>,
-	) -> Result<Quote, Error> {
-		let endpoint = format!("{}/getQuote", CONVERT_API);
-		let params = request_quote_params.try_into_params().map_err(|e| {
-			error!("Error converting request quote params: {}", e);
-			Error::InvalidParams
-		})?;
-		self.make_signed_request(&endpoint, Method::POST, Some(params), recv_window)
-			.await
-	}
-
-	/// Accept the offered quote by quote ID.
-	pub async fn accept_quote(
-		&self,
-		quote_id: &str,
-		recv_window: Option<u32>,
-	) -> Result<ConvertOrder, Error> {
-		let endpoint = format!("{}/acceptQuote", CONVERT_API);
-		let mut params = HashMap::new();
-		params.insert("quoteId".to_string(), quote_id.to_string());
-		self.make_signed_request(&endpoint, Method::POST, Some(params), recv_window)
-			.await
-	}
-
-	/// Get Convert Trade History
-	pub async fn get_convert_trade_history(
-		&self,
-		start_time: u64,
-		end_time: u64,
-		limit: Option<u16>,
-		recv_window: Option<u32>,
-	) -> Result<ConvertTradeHistory, Error> {
-		if let Some(limit) = limit {
-			if limit > MAX_HISTORY_LIMIT {
-				return Err(Error::LimitExceeded);
-			}
-		}
-		// The max interval between startTime and endTime is 30 days.
-		if end_time - start_time > 30 * 24 * 60 * 60 * 1000 {
-			return Err(Error::InvalidParams);
-		}
-
-		let mut params = HashMap::new();
-		params.insert("startTime".to_string(), start_time.to_string());
-		params.insert("endTime".to_string(), end_time.to_string());
-		if let Some(limit) = limit {
-			params.insert("limit".to_string(), limit.to_string());
-		}
-
-		let endpoint = format!("{}/tradeFlow", CONVERT_API);
-		self.make_signed_request(&endpoint, Method::GET, Some(params), recv_window)
-			.await
-	}
-
-	/// Query order status by order ID.
-	pub async fn get_order_status(
-		&self,
-		order_id: Option<String>,
-		quote_id: Option<String>,
-		recv_window: Option<u32>,
-	) -> Result<ConvertOrderStatus, Error> {
-		if order_id.is_none() && quote_id.is_none() {
-			return Err(Error::InvalidParams);
-		}
-		let mut params = HashMap::new();
-
-		if let Some(order_id) = order_id {
-			params.insert("orderId".to_string(), order_id);
-		} else if let Some(quote_id) = quote_id {
-			params.insert("quoteId".to_string(), quote_id);
-		}
-
-		let endpoint = format!("{}/orderStatus", CONVERT_API);
-		self.make_signed_request(&endpoint, Method::GET, Some(params), recv_window)
-			.await
-	}
-
-	/// Enable users to place a limit order
-	pub async fn place_limit_order(
-		&self,
-		place_limit_order_params: PlaceLimitOrderParams,
-		recv_window: Option<u32>,
-	) -> Result<LimitOrder, Error> {
-		let endpoint = format!("{}/limit/placeOrder", CONVERT_API);
-		let params = place_limit_order_params.try_into_params().map_err(|e| {
-			error!("Error converting place limit order params: {}", e);
-			Error::InvalidParams
-		})?;
-		self.make_signed_request(&endpoint, Method::POST, Some(params), recv_window)
-			.await
-	}
-
-	/// Enable users to cancel a limit order
-	pub async fn cancel_limit_order(
-		&self,
-		order_id: String,
-		recv_window: Option<u32>,
-	) -> Result<CancelOrderLimit, Error> {
-		let endpoint = format!("{}/limit/cancelOrder", CONVERT_API);
-		let mut params = HashMap::new();
-		params.insert("orderId".to_string(), order_id.to_string());
-		self.make_signed_request(&endpoint, Method::POST, Some(params), recv_window)
-			.await
-	}
-
-	/// Query open orders
-	pub async fn get_open_orders(&self, recv_window: Option<u32>) -> Result<OpenOrders, Error> {
-		let endpoint = format!("{}/limit/queryOpenOrders", CONVERT_API);
-		self.make_signed_request(&endpoint, Method::POST, None, recv_window).await
-	}
-}
-
-impl BinanceApi {
 	/// Create HMAC SHA256 signature for request parameters
-	fn sign_request(&self, query_string: &str) -> String {
+	pub fn sign_request(&self, query_string: &str) -> String {
 		let mut mac = Hmac::<Sha256>::new_from_slice(self.api_secret.as_bytes())
 			.expect("HMAC can take key of any size");
 		mac.update(query_string.as_bytes());
@@ -197,8 +49,33 @@ impl BinanceApi {
 		hex::encode(code_bytes)
 	}
 
+	/// Helper for making public API requests
+	pub async fn make_public_get_request<T>(
+		&self,
+		endpoint: &str,
+		parameters: Option<HashMap<String, String>>,
+	) -> Result<T, Error>
+	where
+		T: serde::de::DeserializeOwned,
+	{
+		let mut url = self.base_url.join(endpoint).unwrap();
+		if let Some(p) = parameters {
+			url.query_pairs_mut().extend_pairs(p.iter());
+		}
+
+		let response = self.client.get(url.as_str()).send().await.map_err(|e| {
+			error!("API request failed: {}", e);
+			Error::RequestFailed
+		})?;
+
+		response.json::<T>().await.map_err(|e| {
+			error!("Error parsing response: {}", e);
+			Error::ParseResponseFailed
+		})
+	}
+
 	/// Helper for making signed API requests
-	async fn make_signed_request<T>(
+	pub async fn make_signed_request<T>(
 		&self,
 		endpoint: &str,
 		method: Method,
